refactor(portfolio): :sparkles: refactor code
