+++ .babelrc
@@ -1,3 +1,3 @@
 {
-  "presets": ["next/babel", "@zeit/next-typescript/babel"]
+  "presets": ["next/babel"]
 }
