+++ src/components/RoundedImage/index.tsx
