+++ src/layouts/Layout.tsx
@@ -1,7 +1,6 @@
 import React from 'react'
 
-import Navbar from '../components/NavBar/NavBar'
-
+import Navbar from '../components/NavBar'
 const Layout = (props) => {
   // eslint-disable-next-line
   const { children } = props
