+++ src/pages/ponencias.tsx
@@ -3,8 +3,8 @@ import React, { useRef } from 'react'
 import usePaginationArray from '../hooks/usePaginateArray'
 import useInfiniteScroll from '../hooks/useInfiniteScroll'
 import talks from '../assets/data/talks'
-import Card from '../components/Card/Card'
-import ButtonScrollToTop from '../components/ButtonScrollToTop/ButtonScrollToTop'
+import Card from '../components/Card'
+import ButtonScrollToTop from '../components/ButtonScrollToTop'
 
 const Talks = (): JSX.Element => {
   const { dataDisplayed, next, currentPage, maxPage } = usePaginationArray(
